+++ src/modules/index.ts
@@ -33,6 +33,7 @@ export * from './grid-module';
 export * from './badge-of-honor-module';
 export * from './vacuum-module';
 export * from './people-module';
+export * from './navigation-module';
 
 // Export convenience functions
 export { getModuleRegistry } from './module-registry';